+++ build_install_zip.sh
@@ -11,6 +11,9 @@ mkdir -p "$BUILD/site" "$BUILD/admin" "$BUILD/media"
 
 cp "$COMP/com_ordenproduccion.xml" "$BUILD/"
 test -f "$COMP/script.php" && cp "$COMP/script.php" "$BUILD/"
+if [[ -d "$COMP/public" ]]; then
+  cp -R "$COMP/public" "$BUILD/"
+fi
 cp "$COMP/ordenproduccion.php" "$BUILD/site/"
 test -f "$COMP/VERSION" && cp "$COMP/VERSION" "$BUILD/site/"
 cp -R "$COMP/src" "$COMP/tmpl" "$COMP/language" "$COMP/forms" "$BUILD/site/"